Allow per-tenant settings by recreating Settings on switch!
This is a bit more complicated due to the fact that the config gem
removes and sets the Settings constant when reloading from files. This
isn't thread safe as all threads share the same constant. I was also
seeing one thread trying to access Settings when another had just
removed it leading to NameErrors.

My approach to dealing with this is to remove the Settings constant in
an initializer then access the configuration through a Settings
singleton. The Settings class delegates all calls to configuration
stored in the current thread which is created avoiding setting the
Settings constant.

Duplicate of work done in https://github.com/samvera/hyku/pull/1713 so
this can be removed once that PR has been merged and we've updated to
use that version of hyku.