Use both id and source_identifier for imports and exports (#306)
Use both id and source_identifier for imports and exports
If you want to import something that already has a UUID then put it in
both id and source_identifier.  If you want Hyku to mint an id for you
then only put an id in source_identifier.  Exports should contain both
id and source_identifier.