Flaky Hyrax::DOI::RegisterDOIJob fix (#536)
Hyrax::DOI::RegisterDOIJob regularly fails because the cname attribute of account has not been set. Since we are not particularly concerned about this being randomised for this test, we can fix it first and create an account with a matching cname. Then the cname will always have a value and this test hopefully will stop failing.

Remove double definition of account
Fix cname value
Refactor test for brevity
Refactor factories