Fix account behaviour spec failure (#540)
This is a possible fix for the flaky account_behaviour_spec.

In an old commit the behaviour of .reset! was changed so that the mode is reset to :test rather than nil. When switch (not switch! with a bang) is called and fails, then reset! is called (see source code here). Hence the value of mode we should test for in the it "resets the active connections back to the defaults" test will always be :test not nil.

Also it is worth noting that the config specifies the two valid values of mode are symbols not strings, but the test suite uses strings to setup the test objects, and so I have standardised this.